AndroidCourse_SunshinApp
Removed the addLocation from Fetch Weather Task which is no longer
needed. (It's in the parser now.)
